+++ python/GafferVDBUI/__init__.py
@@ -52,5 +52,6 @@ from . import LevelSetOffsetUI
 from . import PointsGridToPointsUI
 from . import SphereLevelSetUI
 from . import PointsToLevelSetUI
+from . import VolumeScatterUI
 
 __import__( "IECore" ).loadConfig( "GAFFER_STARTUP_PATHS", subdirectory = "GafferVDBUI" )